requirements.txt
streamlit
pandas
scikit-learn
pydantic
python-dotenv
huggingface_hub
plotly
streamlit-mic-recorder
